backend/requirements.txt
django 
djangorestframework 
drf-yasg
dotenv
djangorestframework-simplejwt